frozen_requirements.txt
adal==1.0.2
applicationinsights==0.11.5
argcomplete==1.9.4
asn1crypto==0.23.0
azure==3.0.0
azure-batch==4.1.3
azure-cis-scanner==0.2
azure-cli-core==2.0.41
azure-cli-nspkg==3.0.2
azure-common==1.1.13
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.4
azure-datalake-store==0.0.22
azure-eventgrid==0.1.0
azure-graphrbac==0.40.0
azure-keyvault==0.3.7
azure-mgmt==2.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==0.2.0
azure-mgmt-billing==0.1.0
azure-mgmt-cdn==2.0.0
azure-mgmt-cognitiveservices==2.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==3.0.1
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==0.3.1
azure-mgmt-containerregistry==1.0.1
azure-mgmt-containerservice==3.0.1
azure-mgmt-cosmosdb==0.3.1
azure-mgmt-datafactory==0.4.0
azure-mgmt-datalake-analytics==0.3.0
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.3.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==1.2.0
azure-mgmt-eventgrid==0.4.0
azure-mgmt-eventhub==1.2.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iothub==0.4.0
azure-mgmt-iothubprovisioningservices==0.1.0
azure-mgmt-keyvault==0.40.0
azure-mgmt-loganalytics==0.1.0
azure-mgmt-logic==2.1.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementpartner==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==0.2.0
azure-mgmt-monitor==0.4.0
azure-mgmt-msi==0.1.0
azure-mgmt-network==1.7.1
azure-mgmt-notificationhubs==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-powerbiembedded==1.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-recoveryservices==0.2.0
azure-mgmt-recoveryservicesbackup==0.1.1
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.1.0
azure-mgmt-resource==1.2.2
azure-mgmt-scheduler==1.1.3
azure-mgmt-search==1.0.0
azure-mgmt-servermanager==1.2.0
azure-mgmt-servicebus==0.4.0
azure-mgmt-servicefabric==0.1.0
azure-mgmt-sql==0.8.6
azure-mgmt-storage==1.5.0
azure-mgmt-subscription==0.1.0
azure-mgmt-trafficmanager==0.40.0
azure-mgmt-web==0.34.1
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==6.1.2.9
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.3.0
azure-storage-common==1.3.0
azure-storage-file==1.3.0
azure-storage-nspkg==3.0.0
azure-storage-queue==1.3.0
azurerm==0.9.6
backcall==0.1.0
bash-kernel==0.7.1
bcrypt==3.1.4
bleach==2.1.3
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.0.3
cycler==0.10.0
decorator==4.3.0
dominate==2.3.1
entrypoints==0.2.3
Flask==1.0.2
flask-nav==0.6
Frozen-Flask==0.15
futures==3.1.1
html5lib==1.0.1
humanfriendly==4.13
idna==2.6
ipydeps==0.7.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==6.0.1
isodate==0.6.0
itsdangerous==0.24
jedi==0.12.0
jeepney==0.3.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-c-kernel==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter-dashboards==0.7.0
jupyter-nbgallery==0.2.4
keyring==13.2.1
kiwisolver==1.0.1
knack==0.3.3
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
msrest==0.5.2
msrestazure==0.4.34
nbconvert==5.3.1
nbformat==4.4.0
nbstripout==0.3.3
notebook==5.2.2
numpy==1.15.0
oauthlib==2.1.0
ordo==0.3.2
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.4.1
parso==0.2.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.3
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
PyNaCl==1.2.1
pyOpenSSL==17.2.0
pyparsing==2.2.0
pypki2==0.10.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==1.0.0
run-jnb==0.1.12
SecretStorage==3.0.1
simplegeneric==0.8.1
six==1.11.0
tabulate==0.8.2
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==2.0.1
